requirements.txt
PySide6
qtawesome
plotly
pandas
openpyxl
psycopg2-binary